(12/04) Proteção de rotas no front-end e no back-end

+++ back-end/src/controllers/users.js
@@ -6,6 +6,10 @@ const controller = {}; // Objeto vazio
 
 controller.create = async function (req, res) {
     try {
+        // Somente usuários administradores podem acessar este recurso
+        // HTTP 403: Forbidden
+        if (!req?.authUser?.is_admin) return res.status(403).end();
+
         // Verifica se existe o campo "password" e o
         // criptografa antes de criar o novo usuário
         if (req.body.password)
@@ -25,6 +29,10 @@ controller.create = async function (req, res) {
 
 controller.retrieveAll = async function (req, res) {
     try {
+        // Somente usuários administradores podem acessar este recurso
+        // HTTP 403: Forbidden
+        if (!req?.authUser?.is_admin) return res.status(403).end();
+
         // Omite o campo "password" do resultado
         // por questão de segurança
         const result = await prisma.user.findMany({ omit: { password: true } });
@@ -41,6 +49,17 @@ controller.retrieveAll = async function (req, res) {
 
 controller.retrieveOne = async function (req, res) {
     try {
+        // Somente usuários administradores ou o próprio usuário
+        // autenticado podem acessar este recurso
+        // HTTP 403: Forbidden
+        if (
+            !(
+                req?.authUser?.is_admin ||
+                Number(req?.authUser?.id) === Number(req.params.id)
+            )
+        )
+            return res.status(403).end();
+
         const result = await prisma.user.findUnique({
             // Omite o campo "password" do resultado
             // por questão de segurança
@@ -62,6 +81,10 @@ controller.retrieveOne = async function (req, res) {
 
 controller.update = async function (req, res) {
     try {
+        // Somente usuários administradores podem acessar este recurso
+        // HTTP 403: Forbidden
+        if (!req?.authUser?.is_admin) return res.status(403).end();
+
         // Verifica se existe o campo "password" e o
         // criptografa antes de alterar o usuário
         if (req.body.password)
@@ -86,6 +109,10 @@ controller.update = async function (req, res) {
 
 controller.delete = async function (req, res) {
     try {
+        // Somente usuários administradores podem acessar este recurso
+        // HTTP 403: Forbidden
+        if (!req?.authUser?.is_admin) return res.status(403).end();
+
         await prisma.user.delete({
             where: { id: Number(req.params.id) },
         });
@@ -171,6 +198,14 @@ controller.login = async function (req, res) {
         // HTTP 500: Internal Server Error
         res.status(500).end();
     }
+};
+
+controller.logout = function (req, res) {
+    // Apaga no front-end o cookie que armazena o token
+    res.clearCookie(process.env.AUTH_COOKIE_NAME);
+
+    // HTTP 204: No Content
+    res.status(204).end();
 };
 
 controller.me = function (req, res) {
